added launch files for testing base station and raspberry pi computer vision

+++ demining_mqp/ROSNodes/CompVision.py
@@ -17,7 +17,7 @@ class CompVisionNode:
     def searchformine(self, data):
 
         # must be gray-scale image which it should be from the camera
-        gray = data.data  # make this equal to the image from the camera
+        gray = np.uint8(data.data)  # make this equal to the image from the camera
         print gray
         reconstruct = np.zeros((60, 80, 1), dtype=np.uint8)
         print("Array")
